refactor: minor changes across the codebase (#656)
* refactor: move getting add-on version to a separate function

* refactor: get rid of admin_match

* refactor: move GlobalConfigPostProcessor to its own file

* refactor: remove unused method from GlobalConfigPostProcessor

* refactor: rename global_config to global_config_builder_schema

* refactor: no neeed to pass static import_declare_test variable

* chore: get rid of isort
