+++ .pre-commit-config.yaml
@@ -30,10 +30,6 @@ repos:
     rev: 22.12.0
     hooks:
     -   id: black
--   repo: https://github.com/PyCQA/isort
-    rev: 5.11.1
-    hooks:
-    -   id: isort
 -   repo: https://github.com/pre-commit/mirrors-mypy
     rev: v0.991
     hooks:
