+++ pyproject.toml
@@ -54,9 +54,6 @@ importlib-metadata = {version="*", python="<3.8"}
 [tool.poetry.scripts]
 ucc-gen="splunk_add_on_ucc_framework.main:main"
 
-[tool.isort]
-profile = "black"
-
 [build-system]
 requires = ["poetry>=1.0.2"]
 build-backend = "poetry.masonry.api"
